requirements.txt
progressbar2>=3.0
numpy>=1.13
opencv_python>=3.2
scikit-learn>=0.18
scipy>=0.18
matplotlib>=1.5
Pillow>=3.1.2
yacs
tqdm
tensorboardX
pytorch>=1.8.0
torchvision>=0.9.0